+++ R/preflight.R
@@ -162,7 +162,7 @@ preflight_blocking_findings <- function() c(
     "dynamic_dataset_reference", "backward_dependency", "macro_data_flow_deferred",
     "dataset_statement_deferred", "macro_definition_missing",
     "macro_library_initialization_unsupported", "macro_include_requires_expansion",
-    "macro_nested_definition_unsupported")
+    "macro_nested_definition_unsupported", "macro_dependency_analysis_deferred")
 
 preflight_advisory_findings <- function() c("autoexec_autodiscovered", "macro_shadowing",
-  "sasautos_from_environment", "sasautos_from_program")
+  "sasautos_from_environment", "sasautos_from_program", "macro_expansion_unverified")